backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
scikit-learn
xgboost
pandas
numpy
pillow
python-multipart
aiofiles
sqlalchemy
pydantic
geoalchemy2
fastapi
uvicorn[standard]
pillow
torch
torchvision
numpy==1.26.4
sqlalchemy
pydantic
